Don't shout and use Shields.io directly
Of course it's up to you whether you want to make this badge shouty, 
but I don't think it's necessary.

More importantly I'd recommend using the Shields API instead of saving the SVG 
output to your own server. This may end up being hit a lot and we've optimized 
Shields to get hammered by a lot of requests and do the right thing with asset 
caching as much as possible. @espadrine can give you more details if you want.