Create Mission .map and adons for my Bulgarian server
Hello, I maintain a Bulgarian server for your KP LIBERATION missions and we are very pleased with them !!! My request and my question to you and the creators of these missions is the following: Can you make a mission KP LIBERATION on an existing Afghanistan map, but in the menu for buying bots, creating bots to make Bulgarian bots for our Bulgarian server? I will also give you the relevant Bulgarian addons, add-ons to use for bots and equipment against the Taliban. Here are the addons:
https://steamcommunity.com/sharedfiles/filedetails/?id=1622814909
https://steamcommunity.com/sharedfiles/filedetails/?id=1622890915
https://steamcommunity.com/sharedfiles/filedetails/?id=1686005380
Please take my request seriously, think and answer me with understanding and support to our only working server and the Bulgarian fans of your missions KP LIBERATION and the game ARMA 3 !!! ???
Write me an email: matev67@abv.bg
Sincerely yours, konal 777